[feat] add reactive style in FeatureCard
